third_party/WebKit/LayoutTests/fast/dom/HTMLImageElement/image-picture-invalid-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
PASS document.getElementById("canary").clientWidth is 75
PASS document.getElementById("canary").currentSrc is "resources/blue_rect.jpg?0"
PASS document.getElementById("div_end_picture").clientWidth is 800
PASS document.getElementById("div_end_picture").currentSrc is "../../hidpi/resources/image-set-2x.png?2"
PASS document.getElementById("div_multiple_end").clientWidth is 800
PASS document.getElementById("div_multiple_end").currentSrc is "../../hidpi/resources/image-set-2x.png?2"
PASS document.getElementById("source_in_div").clientWidth is 75
PASS document.getElementById("source_in_div").currentSrc is "resources/blue_rect.jpg"
PASS document.getElementById("not_in_picture").clientWidth is 75
PASS document.getElementById("not_in_picture").currentSrc is "resources/blue_rect.jpg"


 
 
 
